# Conflicts:
#	.idea/vcs.xml
#	contracts/libraries/OptionMath.sol
#	contracts/oracle/VolatilityOracle.sol
#	contracts/vault/IVault.sol
#	contracts/vault/VaultRegistry.sol

+++ contracts/vault/IVault.sol
@@ -83,7 +83,7 @@ interface IVault is IERC4626Internal, IERC20Internal {
 
     function getQuote(
         UD60x18 strike,
-        uint64 maturity,
+        uint256 maturity,
         bool isCall,
         UD60x18 size,
         bool isBuy
@@ -91,7 +91,7 @@ interface IVault is IERC4626Internal, IERC20Internal {
 
     function trade(
         UD60x18 strike,
-        uint64 maturity,
+        uint256 maturity,
         bool isCall,
         UD60x18 size,
         bool isBuy,
